Moves game logic from TicTacToeElixir to new Game module
